+++ recipes/_macosx_install.rb
@@ -1,8 +1,6 @@
 # inspiration from https://github.com/caskroom/homebrew-cask/blob/master/Casks/java.rb
 
-cmd = Mixlib::ShellOut.new("pkgutil --pkgs='com.oracle.jdk#{node['java_se']['jdk_version']}'")
-cmd.run_command
-unless cmd.exitstatus == 0
+unless java_version_on_osx?
   version = node['java_se']['version']
 
   name = "JDK #{version.split('.')[1]} Update #{version.sub(/^.*?_(\d+)$/, '\1')}"