+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 7.79.1
+
+- Fix issue with Mac OS X install not using shellout mixlib correctly
+- Use %SYSTEMPATH% on Windows 
+- Use link resource on Windows install
+
 ## 7.79.0
 
 - Branched from 8.60.5 to create 7.79.0